requirements.txt
asgiref==3.3.1
boto3==1.17.39
botocore==1.20.39
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.44.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
environ==1.0
gunicorn==20.1.0
heroku==0.1.4
idna==2.10
jmespath==0.10.0
Markdown==3.3.4
oauthlib==3.1.0
Pillow==8.1.2
pycparser==2.20
PyJWT==2.0.1
python-dateutil==1.5
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.6
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0
